requirements.txt
# Microsoft Agent Framework
agent-framework==1.0.0b251007
agent-framework-devui==1.0.0b251007

# Azure Management SDKs
azure-mgmt-resource
azure-mgmt-storage
azure-mgmt-search
azure-identity

# Azure Services
azure-search-documents==11.7.0b1  # Preview version for Knowledge Agent APIs
azure-storage-blob
openai

# Required dependencies
typing-extensions
python-dotenv
openpyxl  # For Excel file extraction

# Document extraction
azure-ai-documentintelligence>=1.0.0  # Azure Document Intelligence for PDF extraction
python-oxmsg>=0.0.1  # MIT-licensed .msg file parsing (replaces GPL extract-msg)

# RAG pipeline dependencies
langchain-text-splitters  # Semantic chunking with markdown support
tiktoken  # Token counting for chunking

# Evaluation
azure-ai-evaluation  # Azure AI Foundry evaluation SDK

# Testing
pytest
pytest-asyncio  # For testing async functions
